[#8] refactor : add modal molcules and switch all separate  modal functions into it
 - 각 컴포넌트들에 별도로 존재했던 모달컴포넌트를 styles 폴더 내 모달 컴포넌트로 대체
